requirements.txt
langchain
langchain-community
langchain-google-genai
langgraph
fastapi
uvicorn
python-dotenv
requests
pydantic
pydantic-settings
chromadb
pypdf
canvasapi
python-multipart
